Added two new classes: Canvas and IterableTickGenerator.
Canvas is a wrapper for a Surface and its associated
DirectShapeRenderer.  Eventually, most important methods of these two
classes should become part of Canvas, but for now this just holds a
reference to a Surface and a DirectShapeRenderer.

IterableTickGenerator is a wrapper for a TickGenerator (in the
GraphAxis class) that supports iteration.