Added the new class DataPlot.
The DataPlot class is the visual representation of a series of
GrapherTile objects, which is able to draw the data representation on a
GraphWidget object.  The DataPlot class is unfinished at the moment.

Updated documentation in Basis, GraphAxis, GraphWidget, and GrapherTile.
Also updated GraphWidget to use enhanced for loops over lists.