Hello World of Graph Widget.
Added Ajax calls for DataPlot and got them to work.  Also changed
Grapher2 to build a widget with the sine wave graph.

Unfortunately, this does not yet support dragging except on the axes,
and there are three bugs that need to be worked out: when you click on
the Y-axis of this demo, the range immediately goes to [0, 2] rather
than [-1, 1].  Also, we need to find a way to pull tiles so there is
no blank space on the graph, which will probably require changing
currentData in DataPlot to be a list of tiles to draw now, rather than
a single tile.  Finally, the sine wave draws as a horizontal line
for some unknown reason, which might be related to the project2D method
in GraphWidget, although that is pure speculation.