Fixed the "graph draws over the axes" error.
The paintAllDataPoints method in DataPlot now checks the bounds on
points it draws, which ensures that the only data being drawn is in
bounds.

Fixed the "sine graph is a horizontal line" error.  This was actually
from the tiles.  Averaging one sine wave per second over a day, we got
a horizontal line from the server.  This was fixed by reducing the range
(in Grapher2.java) to a value at which the waves are visible.