Updated the handling of pending tiles in DataPlot.
Added a new TileDescription class, which provides convenience methods
dealing with a tile's level and offset.  The DataPlot class now keeps a
set of these objects, and now can keep track of exactly which data is
pending.

Also, switched the PlottablePoint class to immutable, at least in Java
(note that no class is immutable in JavaScript).