Added zoom capabilities to the data view panel.
The current zoom implementation takes any mouse zoom not on the axes
(that is, in the area where data is graphes) and handles a Y-axis zoom
based on that.  This code could easily be modified (just add a loop for
the X-axes in handleMouseWheelEvent of GraphWidget) to zoom on the
X-axes as well, but such a user interface design seems less useful than
one that zooms only on the Y-axes.