Added the getDataPoints method to DataPlot.
This is to make inheritance easier, by allowing subclasses to specify
custom methods to retrieve a List of PlottablePoint objects from a
GrapherTile.  For instance, if the Zeo tiles do not use fields "time"
and "mean", we will be forced to override this data in a ZeoDataPlot
class in order to use the correct fields from the tiles rather than
those the GrapherTile#getDataPoints() method uses.

Also updated some documentation for DataPlot.