Added a ZeoDataPlot class.
This class works mainly by overriding the paintDataPoint method of
DataPlot.  However, we are currently unsure as to the tile format, so
we are not sure if we have to override getDataPoints or not.

Furthermore, we are not sure about how this will draw, since we have not
yet tested Surface#setGlobalAlpha.  As such, we wrote both a G2D and a
native JavaScript implementation, and commented out the call to the
native implementation.  If Surface#setGlobalAlpha does not work as
intended, we can always assign an ID to the GraphWidget object and
then use the native JavaScript implementation.