Added Windows-style zooming on all Chrome instances.
As indicated in the JavaDoc comments, Google Chrome 8 seems to zoom
Windows-style on the Mac.  Since Chrome updates itself very frequently,
almost all Chrome users should be using Chrome 8 or later by the time
this widget is available to the outside world.  Thus, we simply check
for Google Chrome, rather than checking for Chrome 8 or later.