Added an InfoPublisher class to publish information about the Grapher.
This class is not used yet, but at least it offers an interface to
publish information about the Grapher to the rest of the page.

One issue we must deal with is making sure the min/max values for the
axes reflect actual user-visible min/max values, not pixel values.