Added callbacks on tile load.
These callbacks are currently redrawing the grapher every time a tile
loads, which makes for a much smoother user experience.

Oddly enough, in one case with a graph in two parts, though, only
the left half is visibly immediately: the remaining graph is visible
only after the user mouses over the grapher widget.  I am not sure why
this is.  See URL http://localhost:3000/users/4/grapher/datafiles/1
on my machine (that is, the grapher generated from
tilog-anne-100223-fw1.log) for an example of this phenomenon.