Mostly implemented the loading message API in DataPlot.
The only thing left to do is to, as explained in a TODO comment, convert
the representation to Map<String, List<Integer>> so that we can handle
multiple requests for the same URL.  Multiple requests for the same URL
should never happen, but we should make sure that we are safe anyway.

Also improved the loading experience.  Now all but the last tile loads
when the grapher does, even without user interaction to cause the
grapher to paint itself.