Began refactoring the public interface to support photos.
Added the PhotoDescription class, a JavaScript wrapper type that holds
information for a photo.  Also added the PhotoDataPlot class, which is
right now a skeleton but will eventually become a way to embed photos
into the grapher.

Also updated the GrapherTile class to use the PhotoDescription class if
it receives photo data.  I will probably overhaul this class to be no
longer a JavaScript wrapper (and instead hold a reference to a
JavaScript wrapper), but I have not yet done that.  Regardless, changes
to the public interface of the GrapherTile class will probably be
minimal.  The real refactoring is going to come with characterizing a
photo as a PlottablePoint, which it patently is not, at least for the
current PlottablePoint class.