Huge refactoring of the GrapherTile class.
This new version is no longer an overlay type, but more of a wrapper
type for either a PlottablePointTile (the old GrapherTile) or a list of
PhotoDescription objects.

The new interface breaks backward compatibility, but this is OK because
all development so far is internal.  I only had to make a few changes in
DataPlot and ZeoDataPlot to get everything to compile.

The PlottablePointTile class implements all the actions of the old
GrapherTile class, before photos entered in.  Photos are now completely
separated to the PhotoDescription class.

This is a large experimental change - I have not yet tested this code in
a browser.