Cleaned up the interface between GrapherTile and DataPlot.
The GrapherTile interface again breaks backward compatibility, but this
time for the sake of more functionality.  Now a GrapherTile holds its
level and offset, which allows us to determine where to find photo
tiles.  Note that photo tiles do not provide their own level and offset
information, so the only way to place the tiles correctly is to keep
this information available.