Added a new PhotoGetter class.
This class provides a layer of abstraction for image loading, allowing
other classes to avoid worrying about the details of loading.  This
class has a constructor and one method - the getElement() method returns
the requested image if it has loaded yet, and a blank image otherwise.
This makes it easy for another class to show data only if it is
available.