Fixed the bug in which multiple copies of a photo could be drawn.
In the old code, two photos with the same number of seconds t since the
epoch would each get drawn twice, once for each photo in the value
of images.get(new PlottablePoint(t, 0)).  This fixes that by keeping a
set of already-drawn images and only drawing an image once.