Rewrote PhotoGetter as a JavaScript overlay type.
Now I am using the default JavaScript Image object, and the images are
actually drawing correctly.  I still need to add these classes to the
API for redrawing the grapher on load, implement cropping so
images do not cover up the axes, and implement some kind of "highlight
current image" feature, just to get to 0.1.  However, this is a big
difference from the Java-based API that didn't ever fire an onload
event.

Also, I added code in GraphWidget and Grapher2 to define and set the ID
attribute for the Canvas the grapher uses.  This is a bit of a hack that
will need to be changed in the future.