First pass at images moving to accomodate other images.
For some reason, the overlapsLeft and overlapsRight methods in
PhotoDataPlot are never returning true, so this version works exactly
the same as the version from the previous commit.  I need to check this
out more directly.

This version checks for overlap on every draw.  This takes some time, I
am sure, but doesn't seem to affect performance when I try it.  I left a
TODO in PhotoDataPlot to precompute these values whenever I can get the