Fixed the bug that caused the previous version to show no overlap.
The getTime method in PhotoGetter never returned anything, which caused
the error.  This has been fixed.

That having been said, this version still doesn't look very good.
Instead of a lot of big overlapping pictures, now we are getting a bunch
of little overlapping pictures, which I anticipated when I began this
version.  I am going to have to give each picture an index, and force
even-indexed pictures to the top and odd-indexed pictures to the bottom,
grouping images two high.  I also still need to precompute sets of
overlapping images.