Switched from three to two photos high on overlap.
Thus, the images render significantly larger, and by placing photos at
the top or bottom depending on ID, we get a consistent placement of
photos.

However, I might as well document a bug that has been in the code for
a few commits: highlighting is utterly broken whenever images are shrunk
to stop overlap.  Right now, highlighting just uses the middle of the
photos as the Y-value at which they overlap.  This, of course, is not
always correct.  I have documented this bug in a TODO in the code.