Revamped highlighting in DataPlot to set up for highlighting on axes.
Also, discovered a bug that appears to be related to the
BoundedDrawingBox.drawLine method: the rightmost line segment of any
data series in the grapher inexplicably is drawn (at least in Chrome on
Linux), regardless of whether it is too far right or not.  This problem
did not occur with the software highlighting that was in use until
about three commits ago, which is why I blame the
BoundedDrawingBox.drawLine method.