Updated GrapherTile to use the actual level and offset of loaded tiles
If a standard JSON dictionary tile is loaded, then the GrapherTile
object will set its level and offset to match the actual values in the
tile, since the server will not always return the requested level and
offset.  This fixes a bug in which lines going to the left were drawn
from the rightmost tile in a list, under certain conditions.