Rewrote much of GraphWidget.handleMouseMoveEvent
This fix ensures that the axes drag properly whenever the user holds the
mouse button.  I am still not sure if I 100% like the current
functionality, which drops the current dragging object and starts to
drag something else whenever the mouse hits an axis boundary.

Also fixed a bug in which the rightmost data point got rendered even
over the axes.  The problem was that there was an extra beginPath/stroke
pair of calls around the methods we were calling, which caused extra
drawing.