Restored the previous fast performance of the grapher.
Now only calling clip, beginPath, and stroke once per draw operation in
the paintAllDataPoints method.

However, discovered a bug (most likely not from this commit) in which,
if an axis is dragged all the way to the top of the window, so that the
mouse hits the 5-pixel region of no drawing at the top of the window,
the axes start shifting to the right and down, much as they do in an
exception.  However, Chrome developer tools show no exceptions, so I am
not sure what the problem is.