Fixed a bug that caused grapher to look like exception had been thrown.
The bug was causing the graph axes to shift down and to the right on
every repaint, and making the axes draw in a thick stroke, whenever the
mouse dragged a highlighted axis to the top of the graphing window.  The
problem was that I was checking for out-of-bounds values of highlight
lines, and returning if the values were out of bounds.  This stopped
cleanup code from running, which caused the problem.