Built the entire value message API in GraphWidget.
Also overhauled much of the internals of the old API, changing the
representation, drawing methods, and some names.  Specifically,
added a color field to what was the MessageIdPair class but is now the
DisplayMessage class, along with a few other changes in the drawing
method.

Also commented several methods that previously had no JavaDoc comments
at all.

However, note that this version of the code is untested, since not
enough implementation has been added to DataPlot for anything to show up
yet.