Added value display capability to the DataPlot class.
Each DataPlot keeps a single value message ID, and uses it to display a
value whenever that plot is highlighted.  Almost all the work here is
done in a private method called possiblyDisplayHighlightedValue, which
checks the highlightedPoint value and publishes that value to the
display if appropriate.

Also, increased the number of values the graph widget will display at
once from 3 to 4, since it appears that 4 values could fit on the bottom
right.

One issue remains: the text drawn on the grapher appears thick and
fuzzy.  I am not sure why the text is being drawn like this, and will
have to investigate further.