Added code to highlight the current dot the user is mousing over.
Added a new drawing method for subclasses of DataPlot to use:
paintHighlightedPoint.  This method is called on the highlighted point
AFTER all the points have been drawn.  Thus, the highlighted point is
drawn twice, which is why ZeoDataPlot and PhotoDataPlot overrode that
method with no-op methods.  However, in DataPlot, the
paintHighlightedPoint method draws a circle over the highlighted point.

Also added a new constructor to DataPlot so the old interface to the
outside world was restored.