Rewrote GraphWidget.handleMouseWheelEvent
This new version is more like GraphWidget.handleMouseMoveEvent, and
leads to more consistent behavior.  The biggest difference, other than
simplified code, is that, if the user is dragging an axis without the
mouse being over that axis, scrolling without lifting the mouse button
will zoom that axis, rather than zooming whatever the mouse happens to
be over.