Added a new ChannelNamesWidget class
This class will make up the bottom section of the page and any popup
"Add Channels" menu, in either case showing the names of channels.
Currently, this class does not make a difference to the UI, since it has
not been added to the DOM through the RootPanel, and since there is no
kind of callback framework yet.  However, if a callback framework can be
put together, this should work well.