Added a new ChannelManager class
This new class is intended to take the place of the dataPlots, xAxes,
and yAxes private instance variables in the GraphWidget class, in a
major refactoring that enables cleaning up the GraphWidget code
significantly.  However, there is more to this.  The idea is that all
the panels that allow the user to add and remove plots hold a reference
to a single ChannelManager, and then any change made to any of them
automatically propagates to all of them, since they logically hold a
reference to a single data structure.