Fixed two of the glaring errors referenced in the previous commit.
The first fixed error, specifically the error of not adding all the
channels, was fixed by adding the current channel to the list of names
to use, in the else clause of ChannelNamesWidget.addChannelToWidget.

The second fixed error, specifically the error of not checking the
appropriate boxes, was fixed by explicitly checking for inclusion in
visible and checking or unchecking the box appropriately, through a call
to setValue.