Added a panel that allows saving a view name
Lots of TODOs listed in the top of ViewSwitchWidget, since this
implementation is not complete, let alone commented.  I still need to
build a SavableView class (a JavaScript overlay type) that will
represent the data I need to pass back and forth with the server, and
I need to firm up the APIs.  I also need to keep track of the current
view name at all times.