Almost working way to save view to server
The current version conflicts with Rails, failing because it submits a
POST without a CSRF token (I think Rails uses the term
authenticity_token to mean CSRF token).  I am not sure how to fix this
without switching to GET and running into URL size issues, since there
is no form to provide a CSRF token that can be used to authenticate this
application.