Changed ViewSavePopup to override setVisible instead of show
The problem is that PopupPanel.setPopupPositionAndShow hides the window,
sets position, and then makes the window visible again.  Overriding
setVisible ensures that we only change the focus when the popup panel
is actually visible and when changing the focus will actually work.

Also, removed the keyboard listener from the saveName box, since the
browser automatically shifts focus whenever the user presses tab.