Completely revamped the InfoPublisher interface
Now the window.grapherState global variable is changed to a function,
which the outside world can call if it so desires.  The nice thing about
this setup is that with a few more changes to the Java code I will be
able to stop updating the world every time an axis moves, which should
make the application run more smoothly.  Also, this uses a more
consistent interface to talk to the server or to outside JavaScript.
I know of no JavaScript using the current interface, so there should be
no backwards-compatibility issues either.

The only caveat is that the new interface, while tighter, has no way of
publishing color of a channel.  I plan to add this to the view
specification soon, though.

Note that I have not yet added changes made to other classes, so a
checkout of this particular commit won't even compile.