Removed the channel name from the GraphAxis class
There is no longer any need to publish axis bounds under a channel
name, so the GraphAxis class doesn't keep it around any more.  This
allows removal of a lot of code that was essentially passing around
channel names so that GraphAxis could use that channel name to publish
its bounds.

This version of the widget compiles, and runs cleanly except for a bug
with parsing JSON that seems to be related to JSONParser.parseStrict()