Renamed window.grapherState() to window.getCurrentView()
The new function behaves exactly the same as the old one, as detailed in
the comments on the InfoPublisher class.  However, the new name is more
descriptive, and the getXXX syntax makes the function implementation
explicit and doesn't leave clients guessing about whether the
information is in a function call or a global variable.

There are no current known JavaScript users of the window.grapherState
function, so even though this is a backwards-incompatible breaking
change, it is a backwards-incompatible breaking change to prerelease
software without any clients.  Thus, this change is safe