Added code to Canvas to find the native HTML canvas behind its Surface
This way, it is possible to directly access the native element, which
saves calls to getElementById when drawing photos.  Also, in a changeset
I will soon commit, this allows showing photos without setting the
ID attribute of the grapher's canvas.  This way, the grapher is no
longer completely assuming that it is the only copy of itself on the
page.  The only real place where the grapher is completely assuming
that it is alone is in the explicit interaction with outside JavaScript,
through the window.initializeGrapher and window.getCurrentView
functions.