Changed some current code to use WebDownloader
This shortens the code and centralizes any bugfixes.

However, not all code has been converted to use WebDownloader.  In one
part of the application, we use POST, which is not yet supported by
WebDownloader.  Also, I just haven't converted over much of the code
in other parts of the application.