Finished converting all the native code in DataPlotFactory to Java
That is, converted all the native code except for the single
callInitializeGrapher method, which is now static (and thus more widely
usable within the class), as are many of the other methods that were
static before.  However, the biggest change is adding checks to the
callInitializeGrapher and initializeGrapher methods, so that any code
within the DataPlotFactory class will be sure to get a valid, non-null
JSONObject whenever it calls initializeGrapher.  In case of error, that
object will just be the empty dictionary.