Fixed a bug in equality of points with very large or very small times
The old version converted doubles to longs, which overflows (with
possibly undefined behavior - I am not sure about the GWT spec here) if
the value is out of range.  This version keeps everything as a double
and checks for equality slightly differently, making sure that the
floors of two numbers are within 0.5 of each other.