Reverted change from commit master:dee45b9, and made the rest better
Now PhotoDataPlot has an explicit buildPhotoBaseUrl method that
centralizes the base URL creation.  This is because there is a
possibility of confusion with the standard baseUrl from any other type
of data plot.  Also, the DataPlotFactory class has fixed all the issues
related to this parameter (one was introduced when I thought correct
code was a bug, as seen in the commit message for dee45b9.  Thankfully I
tested and caught that bug quickly).