Fixed bug in which order of removing channel listeners changed results
The bug is in InfoPublisher.java, which handles the JavaScript API for
the grapher.  The bug was exposed from the following usage in the Chrome
JavaScript console (">" is the prompt for things I typed, "..." is the
line continuation in this message, since the lines I typed are wider
than 80 characters):

> window.grapher.addChannelChangeListener(function (deviceName,
...    channelName, added) { alert(0); })
0
> window.grapher.addChannelChangeListener(function (deviceName,
...    channelName, added) { alert(1); })
1
> window.grapher.addChannelChangeListener(function (deviceName,
...    channelName, added) { alert(2); })
2
> window.grapher.addChannelChangeListener(function (deviceName,
...    channelName, added) { alert(3); })
3
> window.grapher.removeChannelChangeListner(0)
undefined
> window.grapher.removeChannelChangeListner(2)

At this point, an exception with message "Index: 3, Size: 3" is thrown.

Also, as a reader may have noticed in the output above, the function
that is supposed to be called removeChannelChangeListener is instead
called removeChannelChangeListner.  I fixed this as well.