Added labels on the right side of the Y-axes
This version has a few problems: the code isn't complete (as liberally
indicated by comments), and the corners show a very small but visible
seam line.  Also, this new drawing exposes the fact that the top Y-axis
is cut off outside the edge of the grapher and needs to be taken in a
bit.