Added the NumberAxis class to JavaScript API
This is the beginning of a very large refactoring, so I simply added the
NumberAxis class to JavaScript and made enough changes to GraphAxis to
match.  However, the new model for a GraphAxis does not match the
expectations of a bunch of old code, so the current application should
have no axes (although I haven't tested to be sure).