First JS API grapher version that actually loads in a browser
This does not yet show anything, since there is nothing to call paint
and ask the Java objects to draw themselves.  However, the next step is
to add the GrapherWidget, which will tie together all the pieces.

By "actually loads in a browser" I mean loading the pieces implemented
so far, able to request tiles, and not generating any exceptions.  So
far all these criteria are met, although again there is nothing being
drawn.