Changed the GraphWidget-Surface relationship from is-a to has-a
This is useful because there is no longer a need to initialize a
GraphWidget and add it to a div - the GraphWidget will automatically add
its surface to a div if given a div name as its parameter.  This is a
prelude to making the Surface visible to JavaScript code.

The changes to BodyTrackWidget were only for compatibility to ensure
continued clean compilation.