Removed the ChannelNamesWidget class
Thanks to git, we can bring this class back at a later time.  However,
it uses the ChannelManager so much that refactoring for the new
ChannelManager API will take a lot of work.  This is on top of the fact
that the ChannelNamesWidget does not feature in any of the current or
future BodyTrack designs.