Fixed zero-size canvas and did some cleanup
Canvas is no longer zero size.  Also removed some dead code (stuff that deals with dragging/zooming on an axis, which isn't relevant anymore since axes are in separate divs).  Still not painting (because the GraphWidget holds the reference to the canvas, and the DataPlots don't know about the GraphWidget), but we'll worry about that later.