Added the concept of styles to the DataPlot
There is now a SeriesPlotRenderer interface, which specifies a single
method render.  This method is a series plot's way to render itself.
Right now the DataPlot class uses the new LineRenderer class, which is
basically a cut-past-cleanup of the DataPlot rendering code.  Soon,
though, the DataPlot will take a style as a parameter, allowing it to
render using an arbitrary style.