Removed the loading text publishing code from DataPlot
This does not remove the loading text or highlighted value publishing
API from GraphWidget, only the commented-out code requesting that values
be shown on the GraphWidget.  Thanks to version control, it will be easy
to bring this back if we ever need it.  However, right now the kinds of
publishing that this code enabled does not feature in any BodyTrack
designs.