Fixed display of comments.
I ended up doing away with the pattern of having two renderers per plot (one for highlight, one for no highlight).  Reason was twofold: 1) we want to move more in the direction of a general rendering engine rather than concrete classes; and 2) the renderer was preserving state, in particular the state of the comment panel, so while one renderer thought the comment panel was visible, the other didn't, which caused the comment panel to remain visible when highlighting ended.