Removed the reliance on the default standard.css from GWT
The default standard.css was stomping on page styles, and was included
by default.  So, I placed a copy of that stylesheet in
war/gwt_standard.css, and modified the project properties to include
that style.  The GWT compiler doesn't copy in styles by default, so I
modified the build.xml file to copy gwt_standard.css into war/grapher2.