Several changes to photo display, partially towards proper viewing
Now all the photos show at full height, and there are (temporary)
logging statements to show the parameters for photo drawing, which is
now done much more fully in software rather than trusting to the browser
to handle clipping.  This logging reveals that the problem causing
photos to disappear at too-high resolution is that the drawPhoto methods
isn't even being called whenever the X-axis is zoomed in too much, not
that there is any problem with drawing at too great of a resolution.
Therefore, the next step is to modify the display strategy of photos to
account for this problem.