Revamped photo drawing
This is not yet finished, since the numbers in the corners of photos
still need to be finished.  However, the basic idea of "draw all photos
at full size, only draw photos that won't overlap" is in place.  Also,
this always loads all photos, rather than only loading those photos that
are likely to be visible.