Fixed the bug that caused no count to be drawn on the rightmost photo
Since we draw the count on the previous photo at the same time that we
draw the current photo, and since the rightmost photo is never a
previous photo, the rightmost photo wasn't getting a count.  This was
fixed by adding an explicit drawCount() call to the afterRender() method
of PhotoRenderingStrategy.