Changed PhotoSeriesPlot to lazy-load its photos
Now PhotoGetter.buildPhotoGetter() (both overloads) accepts a download
parameter that lets the caller decide whether or not to download the
photo immediately.  If this is false, the caller can later call the
download() method on the PhotoGetter in order to download the photo.
The PhotoSeriesPlot class takes advantage of this to make photo loading
completely lazy, not downloading a single photo until the drawing logic
determines that it would display on the screen.  This saves bandwidth
and makes the application a snappier experience for the user.