Revamped DataPointListener to include a trigger argument in callbacks
This new trigger argument, also added to the BodyTrack API
documentation, tells the receiver which kind of event occurred and
allows extra information to pass through.  Right now there is no way for
the PhotoSeriesPlot to pass in the photo ID, which is a real need, but
I will add that to the API soon.