Set up the groundwork for publishing of clicked photo points
Added an onClick method to Plot, and an override implementation in
BaseSeriesPlot.  Also added code to call this method whenever the user
clicks on the PlotContainer.  Furthermore, changed DataPointListener to
have one public method that takes two parameters, rather than three
public methods that are the same except for forwarding a different
TriggerAction to the private handleDataPointUpdate method.