Fixed the count bug and removed logging code
The problem was that some photos are downloaded at higher levels and not
at lower levels, causing the count to increase artificially as the
high-level photos are kept in the images field of PhotoSeriesPlot.  The
original fix was to use a HashSet in order to zero out the counts of
photos that aren't part of the current tile.  However, this failed
because of GWT's custom equals() and hashCode() on classes inheriting
from JavaScriptObject.  So, I wrote my own Set implementation that wraps
a normal HashSet and passes in objects with externally defined equals()
and hashCode(), via some object implementing the new interface
NativeObjectSet.EqualsHashcodeProvider.  This causes the counts to stay
consistent throughout use of a photo channel.