Added guaranteed update whenever the user lets go of the mouse button
Also, whenever the mouse leaves a plot container or graph axis, there is
a guaranteed update.  This removes the possibility of an annoying "jump"
in the channels if the user drags quickly and the last 50 ms of the drag
doesn't get passed in the update.  However, since there is no way to
know which scroll event will be the last, there can still be a lag in
the update between the last scroll event and the update in the plot
containers.