Finished implementing mandatory midnight lines
Now the lines are actually drawn at midnight local time, as on the
X-axis.  Also, the lines are always drawn at least 20 pixels apart.
When the channel is 600 pixels wide (this is a guesstimate on my part of
our 800-pixel minimum for the entire application, minus space for the
channel label and whitespace), 30 days are shown before the lines
disappear.