Completely changed out the implementation of mouse wheel deltas
Now BaseMouseWheelHandler detects the browser and platform to determine
how to scale the native mouse wheel delta, and uses the scaled native
delta instead of GWT's delta.  This has the advantage of being close to
GWT's delta, but a double.  Since this value is a double, slow scrolling
isn't ignored any more.  This certainly works on Chrome 17 and Firefox 11
for Linux, as well as Chrome 17 and Safari 5.0.5 for Mac OS X 10.6.8.
The entire BodyTrack site chokes and dies on Firefox 3.5 for Mac OS X
10.6.8, so I can't really test the scroll behavior.  I still need to
test Safari 5.1 and later for Mac OS X (edit: it appears that Safari 5.1
scrolls nicely under this code), and test Windows browsers (IE 9 and
IE 10, Chrome, Firefox 11, and Safari).