Refactored and reformatted BaseSeriesPlotRenderer.render
Now render just calls three other methods in order, and those methods do
the heavy lifting.

In the process, reordered the iteration between plot renderers and
tiles, so that there are no line breaks between tiles.  However, this
exposes a bug in PhotoSeriesPlot that causes an
IndexOutOfBoundsException, which I need to debug.