Got closer to fixing the grapher crash on nanosecond zoom
This uses the new MathEx.ulp(double) method - since GWT doesn't provide
a Math.ulp method - to calculate the proper epsilon on the fly, rather
than using a fixed epsilon that will work for times near enough to the
epoch but won't work today or in the future.

Also deleted the log2 method in StandardTileLoader, since log2 fits
better in MathEx.