Changed TimeGraphAxis.getOffsetAtTimestamp to ignore timezone end times
The web API to tell the user's timezones has the potential to return
overlapping regions or regions with gaps.  This change removes the check
for the end of a timezone (the only place in the grapher that cares
about the end of a timezone), which allows the grapher to handle these
situations gracefully.

The assumption that a gap in the data means that the user is in the
previous timezone is reasonable, because the user may turn off their
device or accidentally upload wrong data, all without moving.  It is
unlikely that a gap in the data means that the user took a trip to a
different timezone, because most users don't regularly travel to a
different timezone.