No longer downloading full-size if thumnails are present
Previously, PhotoGetter treated the full-size image as the largest
thumbnail, and downloadIfNecessary() chose to download it whenever the
(minHeight, maxHeight) range no longer included any thumbnails.  Now,
PhotoGetter.downloadIfNecessary() never downloads the full-size image if
any thumbnails are available, but always downloads the full-size image
if (a) no thumbnails are available, and (b) the full-size image has not
already been requested.

This policy change greatly simplifies the adjustments for rotation in the
EXIF section of an image file - the server rotates the thumbnails, but
the full-size image in Fluxtream Capture is not rotated.  By using only
thumbnails when any are available, this change obviates the need for a
server change to rotate the full-size image or a client-side change to
rotate the full-size image when displaying it in the timeline.