Added localDisplay support to the grapher (breaking constructor change)
This change switches the last optional parameter passed to the
DataSeriesPlot and PhotoSeriesPlot constructors from a single optional
style to a dictionary of optional parameters.  One of the accepted keys
is "style", which has the same effect as the previous optional
parameter.  This change also adds new key "localDisplay", which treats
the associated value as a boolean with default value false.  If this is
set to true, the data from the server is treated specially, as local
time data.

The motivation for local time display is that some services give a time
but no time zone, which makes the time essentially meaningless.  For
instance, if we know that some data point happened at 8pm, but don't
know the timezone as context for that 8pm number, the previous version
of the grapher can't really handle the data, since the grapher is
entirely based on UTC internally.  The solution to this is to display
the data point so that it matches up with 8pm on the X-axis, even
though that may not match the real time that the data came from.  There
is missing information here, and it's better to display a sunrise photo
under the 7am mark on the time axis than to display it at 1am because
the user's browser is in a different time zone than the photo was taken.

We implement this feature by pretending on server side that the data was
recorded in UTC, even if it wasn't.  Therefore, the server stores the
7am photo at 7am UTC, and we request the tile containing 7am browser
time as though we were requesting a tile for 7am UTC, and then shift the
data to display it at the 7am mark in the timeline.  All this shifting
work is implemented in TimeZoneMap and the classes that derive from it,
and is neatly wrapped using ModalTimeZoneMap.