Fixed grapher constant tile load due to mixed-up variable name
I typed the wrong variable name on one line in ModalTimeZoneMap in the
previous commit, and so returned undefined whenever the channel wasn't
in local time mode.  This got turned into a null data points field,
which caused a NullPointerException on an attempt to iterate over the
data points, which caused a failure handler to execute in the Fluxtream
app that was executing this library code, which caused an
attempt to reload the tile.  This happened over and over and over, as
fast as the server and browser would keep up.

To fix this bug, I just fixed the variable name.